- Reworked urls.py to point to a single router method in views.py.
- Reworked urls in api.js, got rid of 'test' prefix.

+++ server/todolist/todolist/views.py
@@ -10,64 +10,52 @@ from .utils import utils
 
 
 @api_view(['GET', 'POST', 'DELETE'])
-def test(request):
-    if request.method == 'POST':
-        new_list_item_data = request.data
-        utils.persist_new_list_item(new_list_item_data)
+def router(request, idk):
+    request_method = request.method
+    request_path = request.path
+    request_data = request.data
+
+    if 'persist_new_list_item' in request_path:
+        utils.persist_new_list_item(request_data)
 
         return Response(status=status.HTTP_200_OK)
 
-    elif request.method == 'GET':
-        obj = TodoListItem.objects.all()
-        data = serializers.serialize("json", obj, fields="list_item_title")
+    elif 'get_table_data' in request_path:
+        list_name = request.GET.get('listName')
+        obj = TodoListItem.objects.filter(associated_list=list_name)
+        data = serializers.serialize("json", obj, fields=('checked', 'list_item_title'))
 
         return Response(data, status=status.HTTP_200_OK)
 
-    return Response(status=status.HTTP_200_OK)
+    elif 'persist_new_list' in request_path:
+        list_name = request.GET.get('listName')
+        new_list = Lists(list_name=list_name)
+        new_list.save()
 
+        return Response(status=status.HTTP_200_OK)
 
-@api_view(['GET'])
-def persist_new_list(request):
-    list_name = request.GET.get('listName')
-    new_list = Lists(list_name=list_name)
-    new_list.save()
+    elif 'update_checked_item' in request_path:
+        checked_item_id = request.GET.get('checkedItemID')
+        list_item = TodoListItem.objects.get(id=checked_item_id)
+        list_item.checked = not list_item.checked
+        list_item.save()
 
-    return Response(status=status.HTTP_200_OK)
+        return Response(status=status.HTTP_200_OK)
 
+    elif 'get_lists' in request_path:
+        obj = Lists.objects.all()
+        data = serializers.serialize("json", obj, fields='list_name')
 
-@api_view(['GET'])
-def update_checked_item(request):
-    checked_item_id = request.GET.get('checkedItemID')
-    list_item = TodoListItem.objects.get(id=checked_item_id)
-    list_item.checked = not list_item.checked
-    list_item.save()
+        return Response(data, status=status.HTTP_200_OK)
 
-    return Response(status=status.HTTP_200_OK)
+    elif 'delete_list_items' in request_path:
+        items_to_delete = request.data
+        TodoListItem.objects.filter(id__in=items_to_delete).delete()
 
+        return Response(status=status.HTTP_200_OK)
 
-@api_view(['GET'])
-@renderer_classes([JSONRenderer])
-def get_table_data(request):
-    list_name = request.GET.get('listName')
-    obj = TodoListItem.objects.filter(associated_list=list_name)
-    data = serializers.serialize("json", obj, fields=('checked','list_item_title'))
+    elif 'hi' in request_path:
+        return Response(status=status.HTTP_200_OK)
 
-    return Response(data, status=status.HTTP_200_OK)
-
-
-@api_view(['GET'])
-@renderer_classes([JSONRenderer])
-def get_lists(request):
-    obj = Lists.objects.all()
-    data = serializers.serialize("json", obj, fields='list_name')
-
-    return Response(data, status=status.HTTP_200_OK)
-
-
-@api_view(['DELETE'])
-@renderer_classes([JSONRenderer])
-def delete_list_items(request):
-    items_to_delete = request.data
-    TodoListItem.objects.filter(id__in=items_to_delete).delete()
-
-    return Response(status=status.HTTP_200_OK)
+    # if all else fails
+    return Response(status=status.HTTP_404_NOT_FOUND)
